requirements.txt
beautifulsoup4==4.12.2
google==3.0.0
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
soupsieve==2.5
